main:creating central main function dynamodb: relocating dynamodb from main create_api:returning id sendRequest: making url from id
