+++ sendRequest.py
@@ -1,19 +1,21 @@
 import requests
 import json
 
-url = "http://localhost:4566/restapis/2cdvg7tg91/dev/_user_request_/roles"
-data = {
-    "action": "put",
-    "id": "1",
-    "type": "admin",
-    "permissions": {"read": True, "write": True}
-}
 
-response = requests.post(url, headers={"Content-Type": "application/json"}, data=json.dumps(data))
-print(response.status_code)
-print(f"Status code: {response.status_code}")
-print(f"Response text: {response.text}")
-try:
-    print(response.json())
-except Exception:
-    print("Brak JSON-a w odpowiedzi.")
+def send_request(api_id):
+    url = "http://localhost:4566/restapis/" + api_id + "/dev/_user_request_/roles"
+    data = {
+        "action": "put",
+        "id": "1",
+        "type": "admin",
+        "permissions": {"read": True, "write": True}
+    }
+
+    response = requests.post(url, headers={"Content-Type": "application/json"}, data=json.dumps(data))
+    print(response.status_code)
+    print(f"Status code: {response.status_code}")
+    print(f"Response text: {response.text}")
+    try:
+        print(response.json())
+    except Exception:
+        print("Brak JSON-a w odpowiedzi.")
